@@ -1,28 +1,26 @@
 {
-  "name": "badge-generator",
+  "name": "glass-badgic",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "clean": "trash dist",
     "dev": "vite",
-    "build": "npm run clean && tsc && vite build && touch dist/.nojekyll",
-    "deploy": "npm run build && node ./gh-pages.mjs",
+    "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "framer-motion": "^6.2.4",
-    "react": "^17.0.2",
-    "react-colorful": "^5.5.1",
-    "react-dom": "^17.0.2",
-    "styled-components": "^5.3.3"
+    "react": "^18.2.0",
+    "react-colorful": "^5.6.1",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/styled-components": "^5.1.22",
-    "@vitejs/plugin-react": "^1.0.7",
-    "gh-pages": "^3.2.3",
-    "trash-cli": "^5.0.0",
-    "typescript": "^4.4.4",
-    "vite": "^3.0.9"
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react": "^3.1.0",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.2.7",
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0"
   }
 }
